move rest_network_test to agent_test
